Add ApolloProvider to App.tsx and authProvider.tsx
