+++ __tests__/Notifications.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Notifications from '../screens/Notifications';
-import renderWithAuth from '../config/AuthProvider';
+import renderWithAuth from '../utils/authProvider';
 
 describe('Notifications Screen', () => {
   test('renders the Notifications screen', () => {
