+++ __tests__/Server.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react-native';
 import Server from '../screens/Server';
-import renderWithAuth from '../config/AuthProvider';
+import renderWithAuth from '../utils/authProvider';
 
 describe('Server screen', () => {
   test('renders correctly', () => {
