+++ graphql/queries/Collection.ts
@@ -1,10 +1,10 @@
 import { gql } from '@apollo/client';
 
-export const GET_CONTACTS = gql`
+export const GET_COLLECTIONS = gql`
   query search($filter: SearchFilter!, $contactOpts: Opts!, $messageOpts: Opts!) {
     search(filter: $filter, contactOpts: $contactOpts, messageOpts: $messageOpts) {
-      contact {
-        name
+      group {
+        label
       }
     }
   }
