+++ screens/Collections.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
-import { ApolloProvider } from '@apollo/client';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 
-import { client } from '../config/apollo';
 import SearchBar from '../components/ui/SearchBar';
 import CollectionList from '../components/ui/CollectionList';
 
@@ -18,10 +16,8 @@ type Props = NativeStackScreenProps<RootStackParamList, 'Contacts'>;
 const Collections = ({ navigation }: Props) => {
   return (
     <View style={styles.mainContainer}>
-      <ApolloProvider client={client}>
-        <SearchBar />
-        <CollectionList navigation={navigation} />
-      </ApolloProvider>
+      <SearchBar />
+      <CollectionList navigation={navigation} />
     </View>
   );
 };
